Created index and product components, created App.vue and modified app.js
I modified the app js making the imports of the vue, the imports of the components and defined the routes of the index and the products.

I created the App vue that is the index of the page and added it the index and the product link to the component.

I modified the index.blade.php adding the div app to show the components in the public folder to the users.
I created the index component that will be the index page of the website

I created the products component that is the view that will present all of the web site products (it does not showing nothing yet)